+++ cobra/design/design_algorithms.py
@@ -1,4 +1,4 @@
-from cobra.util import DefunctError
+from cobra.exceptions import DefunctError
 
 
 def set_up_optknock(*args, **kwargs):
