+++ cobra/topology/reporter_metabolites.py
@@ -1,6 +1,4 @@
-# Based on Patil et al 2005 PNAS 102:2685-9
-# TODO: Validate cobra.core compliance
-from cobra.util import DefunctError
+from cobra.exceptions import DefunctError
 
 
 def identify_reporter_metabolites(*args, **kwargs):
