+++ cobra/util/__init__.py
@@ -1,7 +1,7 @@
 from cobra.util.util import *
 from cobra.util.context import *
 from cobra.util.solver import *
-from cobra.util.exceptions import *
+
 
 try:
     import numpy
